+++ app/src/main/java/com/app/cellstudio/hacker_news_app/interactor/scheduler/ImmediateSchedulerProvider.kt
@@ -1,0 +1,19 @@
+package com.app.cellstudio.hacker_news_app.interactor.scheduler
+
+import com.app.cellstudio.androidkotlincleanboilerplate.interactor.scheduler.BaseSchedulerProvider
+import io.reactivex.Scheduler
+import io.reactivex.schedulers.Schedulers
+
+class ImmediateSchedulerProvider : BaseSchedulerProvider {
+    override fun computation(): Scheduler {
+        return Schedulers.trampoline()
+    }
+
+    override fun io(): Scheduler {
+        return Schedulers.trampoline()
+    }
+
+    override fun ui(): Scheduler {
+        return Schedulers.trampoline()
+    }
+}