+++ app/src/main/java/com/app/cellstudio/hacker_news_app/interactor/viewmodel/DetailsViewModel.kt
@@ -1,0 +1,8 @@
+package com.app.cellstudio.hacker_news_app.interactor.viewmodel
+
+import com.app.cellstudio.hacker_news_app.interactor.model.PageModel
+import io.reactivex.Observable
+
+interface DetailsViewModel : ViewModel {
+    fun getDetailsPageModel(): Observable<List<PageModel>>
+}