+++ app/src/main/java/com/app/cellstudio/hacker_news_app/interactor/viewmodel/impl/MainViewModelImpl.kt
@@ -1,30 +1,28 @@
 package com.app.cellstudio.hacker_news_app.interactor.viewmodel.impl
 
-import android.databinding.ObservableBoolean
 import com.app.cellstudio.androidkotlincleanboilerplate.interactor.scheduler.BaseSchedulerProvider
 import com.app.cellstudio.hacker_news_app.data.repository.HackerNewsRepository
 import com.app.cellstudio.hacker_news_app.interactor.model.HackerNewsItemModel
 import com.app.cellstudio.hacker_news_app.interactor.viewmodel.BaseViewModel
 import com.app.cellstudio.hacker_news_app.interactor.viewmodel.MainViewModel
 import io.reactivex.Observable
+import io.reactivex.subjects.PublishSubject
 
 class MainViewModelImpl(
     private val hackerNewsRepository: HackerNewsRepository,
     scheduler: BaseSchedulerProvider
 ) : BaseViewModel(scheduler), MainViewModel {
 
-    private val isLoading = ObservableBoolean(false)
+    private val isLoading = PublishSubject.create<Boolean>()
 
-    override fun getLoading(): ObservableBoolean {
+    override fun getLoading(): Observable<Boolean> {
        return isLoading
     }
 
-
-
     override fun getTopStories(): Observable<List<Int>> {
-        isLoading.set(true)
+        isLoading.onNext(true)
         return hackerNewsRepository.getTopStories(LIMIT_TO_START)
-            .doFinally { isLoading.set(false) }
+            .doFinally { isLoading.onNext(false) }
     }
 
     override fun getHackerNewsItemModel(id: Int): Observable<HackerNewsItemModel> {
@@ -32,6 +30,6 @@ class MainViewModelImpl(
     }
 
     companion object {
-        private val LIMIT_TO_START = 25
+        private const val LIMIT_TO_START = 25
     }
 }