+++ app/src/main/java/com/app/cellstudio/hacker_news_app/presentation/adapter/TopStoriesAdapter.kt
@@ -2,20 +2,17 @@ package com.app.cellstudio.hacker_news_app.presentation.adapter
 
 import android.databinding.DataBindingUtil
 import android.support.v7.widget.RecyclerView
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.app.cellstudio.hacker_news_app.R
 import com.app.cellstudio.hacker_news_app.databinding.ListTopStoriesBinding
-import com.app.cellstudio.hacker_news_app.databinding.ListTopStoriesEmptyBinding
 import com.app.cellstudio.hacker_news_app.interactor.model.HackerNewsItemModel
 import io.reactivex.subjects.PublishSubject
 
 class TopStoriesAdapter(private val topStories: MutableList<Int>) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
 
     private var topStoriesData: Array<HackerNewsItemModel?> = arrayOfNulls(topStories.size)
-    private var loading: Boolean = false
     private val selectedModel = PublishSubject.create<HackerNewsItemModel>()
     private var listener: Listener? = null
 
@@ -30,19 +27,10 @@ class TopStoriesAdapter(private val topStories: MutableList<Int>) : RecyclerView
         }
     }
 
-    class EmptyViewHolder : RecyclerView.ViewHolder {
-
-        lateinit var binding: ListTopStoriesEmptyBinding
-
-        constructor(view: View) : super(view)
-
-        constructor(binding: ListTopStoriesEmptyBinding) : this(binding.root) {
-            this.binding = binding
-        }
-    }
+    class EmptyViewHolder(view: View) : RecyclerView.ViewHolder(view)
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-        var viewHolder: RecyclerView.ViewHolder?
+        val viewHolder: RecyclerView.ViewHolder?
         val layoutInflater = LayoutInflater.from(parent.context)
 
         viewHolder = when (viewType) {
@@ -59,18 +47,15 @@ class TopStoriesAdapter(private val topStories: MutableList<Int>) : RecyclerView
     }
 
     override fun onBindViewHolder(baseHolder: RecyclerView.ViewHolder, position: Int) {
-        val pos = baseHolder.adapterPosition
         if (baseHolder is EmptyViewHolder) {
-            Log.d("CallApi", pos.toString())
-            listener?.onFirstVisible(topStories[pos], pos)
+            // Call API if is empty view holder
+            listener?.onFirstVisible(topStories[position], position)
         } else if (baseHolder is ViewHolder) {
-            if (pos >= 0) {
-                baseHolder.binding.modelId = "No. " + (position + 1).toString()
-                baseHolder.binding.model = topStoriesData[position]
-                baseHolder.binding.setListener {
-                    if (topStoriesData[position] != null) {
-                        selectedModel.onNext(topStoriesData[position]!!)
-                    }
+            baseHolder.binding.modelId = "No. " + (position + 1).toString()
+            baseHolder.binding.model = topStoriesData[position]
+            baseHolder.binding.setListener {
+                if (topStoriesData[position] != null) {
+                    selectedModel.onNext(topStoriesData[position]!!)
                 }
             }
         }
@@ -97,6 +82,13 @@ class TopStoriesAdapter(private val topStories: MutableList<Int>) : RecyclerView
     fun updateViewHolder(position: Int, hackerNewsItemModel: HackerNewsItemModel) {
         topStoriesData[position] = hackerNewsItemModel
         notifyItemChanged(position)
+    }
+
+    fun updateData(models: List<Int>) {
+        this.topStoriesData = arrayOfNulls(models.size)
+        this.topStories.clear()
+        this.topStories.addAll(models)
+        notifyDataSetChanged()
     }
 
     companion object {