+++ RELEASENOTES.md
@@ -7,3 +7,14 @@
 * Setup MVVM architecture
 * Setup Main Page
 
+## 1.1.0
+
+### Features
+
+* Setup Details Page
+* Setup Article Fragment and Comment Fragment
+* Add in swipe to refresh feature in Main page
+
+### Bug Fixes
+* Remove Flexbox layout
+* Fix Portrait Orientation